Handle case when Black or Flake8 aren't available
Before this change, if Flake8 wasn't installed (or otherwise not
available) there would be no user feedback and the issues sidebar would
say "No issues found.". It was even worse for Black: it would remove the
text in the editor window and you'd have to use Undo to get it back.

Now, when either Flake8 or Black isn't available a notification is sent
and the unformatted code is left in place.