Refactor for asynchronous saving
When saving a document, Nova would save the file, then Black would
apply the formatting after the save. This left the document as unsaved
_after an explicit save_. To ensure that nova waits for the Black
process to finish, Nova's onWillSave method needs to return a Promise.
It will wait for 5 seconds while that promise resolves, ensuring that
the document will be reformatted first and saved second. This refactor
attempts just that.

The main refactor is around the execute method of BlackProcess and the
format method of Formatter. execute now returns a promise. If the
promise resolves, it passes the formatted text to the formatter, which
calls the edit function on the document to make the change. If the
promise rejects it represents a failure in some way. The formatter
catches this rejection and creates a notification for the user. It's
often (but not always) the case that the error is something the user
would want to know about, like a malformed python document.

This also

* Formatted the edited files through Prettier
* removed `didExit` from BlackProcess
* Removed some unused async declarations from BlackProcess
* Formatter returns the results of BlackProcess.execute
* main returns the result of Formatter.format